Complete login system using JWT
Complete login system using a created method in the user model file to create a JWT token with a secret key I have generated using Nord pass password manager. I will generate a new key in the future and store it in a more secure manner such as .env file. |