Minor errors fixed for video model + validation
Validation for video uploading needs the date to be a NodeJS date object, I simply created a temp variable to convert the received date from API to a date object and then test it.

Also minor error where stu_email was of type number and not String in the model.